Fix clang build
The repo now builds correctly with the latest clang & libc++.

Two main problems are addressed with this change:

1. Modification of an inactive union member is not allowed in constexpr, and this
causes clang to disallow member assignment. However, the whole union object can
be assigned with a new one. This could be a bug in the standard wording, or of
either GCC's or Clang's interpretation, but clearly assignment is useful and
probably not what is intended to be prohibited?

As part of the union change, the various ExternalView members have been
collapsed into one.

2. Clang cannot correctly infer the return type of the mutually recursive functions
that return parsers as lambdas. We can get around this by:
 * providing help in the form of decay-to-function-pointer where that is
 possible
 * making our own lambda closure type where decay-to-function-pointer is not
   possible

Other minor problem fixed here: clang complains about control reaching the end of a
non-void function in the value indexing operators.